requirements.txt
gensim
matplotlib
scikit-learn
transformers
pillow
torch
timm
